+++ cplusplus/include/instructions/InstructionBase.hpp
@@ -152,6 +152,9 @@ namespace rabbitizer {
         TrinaryValue FlagGet_disasmAsData() const;
         void FlagSet_disasmAsData(TrinaryValue value);
 
+        TrinaryValue FlagGet_r5900UseDollar() const;
+        void FlagSet_r5900UseDollar(TrinaryValue value);
+
         /* flags */