+++ include/instructions/RabbitizerInstruction.h
@@ -38,12 +38,16 @@ typedef struct RabbitizerInstruction {
      * Flags are bitpacked, refer to the `RAB_INSTR_FLAGS_` macros to access them.
      *
      * Bit usage:
-     * - Bits 0 ~ 1: Value of the `RabTrinaryValue` enum.
+     * - Bits 0 ~ 1: `disasmAsData`. Value of the `RabTrinaryValue` enum.
      *     - `RAB_TRINARY_VAL_TRUE` forces the instruction to be disassembled as data.
      *     - `RAB_TRINARY_VAL_FALSE` bypasses the global checks for disassembling a word as data. A word will still be disassembled as data if it can't be decoded.
      *     - `RAB_TRINARY_VAL_NONE` leaves this decision to the global settings.
      *     - Defaults to `RAB_TRINARY_VAL_NONE`
-     * - Bit 2 ~ 3: Reserved for future use.
+     * - Bit 2 ~ 3: `r5900UseDollar`. Value of the `RabTrinaryValue` enum.
+     *     - `RAB_TRINARY_VAL_TRUE` forces the use of dollar signs ($) on R5900's VU instructions
+     *     - `RAB_TRINARY_VAL_FALSE` forces disassembling to not use of dollar signs ($) on R5900's VU instructions
+     *     - `RAB_TRINARY_VAL_NONE` leaves this decision to the global settings.
+     *     - Defaults to `RAB_TRINARY_VAL_NONE`
      * - Bit 4 ~ 5: Reserved for future use.
      * - Bit 6 ~ 7: Reserved for future use.
      * - Bit 8 ~ 9: Reserved for future use.
@@ -143,6 +147,9 @@ typedef struct RabbitizerInstruction {
 
 #define RAB_INSTR_FLAGS_GET_disasmAsData(self)                            (RabTrinaryValue)(SHIFTR((self)->flags,  0,  2))
 #define RAB_INSTR_FLAGS_SET_disasmAsData(self, value)          ((self)->flags = BITREPACK((self)->flags, (value),  0,  2))
+
+#define RAB_INSTR_FLAGS_GET_r5900UseDollar(self)                          (RabTrinaryValue)(SHIFTR((self)->flags,  2,  2))
+#define RAB_INSTR_FLAGS_SET_r5900UseDollar(self, value)        ((self)->flags = BITREPACK((self)->flags, (value),  2,  2))
 
 
 NON_NULL(1)